refactor the lawyer dashboard
