+++ src/components/dashboards/features/lawyer/WelcomeSection.js
@@ -1,0 +1,10 @@
+import React from 'react';
+
+const WelcomeSection = () => (
+  <div>
+    <h2 className="text-2xl font-semibold text-primary-light mb-4">Welcome</h2>
+    <p className="text-secondary-light">Please select an option from the sidebar.</p>
+  </div>
+);
+
+export default WelcomeSection;
